// src/app/domains/shared/models/preliquidacion.model.ts
export interface PreLiquidacion {
    idRuta:number;
    metaRecaudo: number;
    baseEntregada: number;
    totalRecaudado: number;
    totalGastos: number;
    totalPrestamosNuevos: number;
    realEntregar:number;
    entregaEfectivo:number;
    faltanteGenerado:number;
    sobranteGenerado:number
  }
  